feat(cb2-10603): clean up
