+++ tests/unit/reportGenFunction.unitTest.ts
@@ -1,3 +1,5 @@
+const mockProcessRecord = jest.fn();
+
 import { reportGen } from "../../src/functions/reportGen";
 import { ReportGenerationService } from "../../src/services/ReportGenerationService";
 import { SendATFReport } from "../../src/services/SendATFReport";
@@ -5,8 +7,9 @@ import { SendATFReport } from "../../src/services/SendATFReport";
 
 jest.mock("../../src/services/ReportGenerationService");
 jest.mock("../../src/services/SendATFReport");
-
-const mockPayload = "{\n  \"Type\" : \"Notification\",\n  \"MessageId\" : \"-c9d5---\",\n  \"TopicArn\" : \"tf-visit\",\n  \"Message\" : \"{\\\"eventID\\\":\\\"f9e63bf29bd6adf174e308201a97259f\\\",\\\"eventName\\\":\\\"MODIFY\\\",\\\"eventVersion\\\":\\\"1.1\\\",\\\"eventSource\\\":\\\"aws:dynamodb\\\",\\\"awsRegion\\\":\\\"eu-west-1\\\",\\\"dynamodb\\\":{\\\"ApproximateCreationDateTime\\\":1711549645,\\\"Keys\\\":{\\\"id\\\":{\\\"S\\\":\\\"6e4bd304-446e-4678-8289-dasdasjkl\\\"}},\\\"NewImage\\\":{\\\"testerStaffId\\\":{\\\"S\\\":\\\"132\\\"},\\\"testStationPNumber\\\":{\\\"S\\\":\\\"87-1369564\\\"},\\\"testerEmail\\\":{\\\"S\\\":\\\"tester@dvsa.gov.uk1111\\\"},\\\"testStationType\\\":{\\\"S\\\":\\\"gvts\\\"},\\\"testStationEmail\\\":{\\\"S\\\":\\\"teststationname@dvsa.gov.uk\\\"},\\\"startTime\\\":{\\\"S\\\":\\\"2022-01-01T10:00:40.561Z\\\"},\\\"endTime\\\":{\\\"S\\\":\\\"2022-01-01T10:00:40.561Z\\\"},\\\"id\\\":{\\\"S\\\":\\\"6e4bd304-446e-4678-8289-dasdasjkl\\\"},\\\"testStationName\\\":{\\\"S\\\":\\\"Rowe, Wunsch and Wisoky\\\"},\\\"activityType\\\":{\\\"S\\\":\\\"visit\\\"},\\\"activityDay\\\":{\\\"S\\\":\\\"2022-01-01\\\"},\\\"testerName\\\":{\\\"S\\\":\\\"namey mcname\\\"}},\\\"OldImage\\\":{\\\"testerStaffId\\\":{\\\"S\\\":\\\"132\\\"},\\\"testStationPNumber\\\":{\\\"S\\\":\\\"87-1369564\\\"},\\\"testerEmail\\\":{\\\"S\\\":\\\"tester@dvsa.gov.uk1111\\\"},\\\"testStationType\\\":{\\\"S\\\":\\\"gvts\\\"},\\\"testStationEmail\\\":{\\\"S\\\":\\\"teststationname@dvsa.gov.uk\\\"},\\\"startTime\\\":{\\\"S\\\":\\\"2022-01-01T10:00:40.561Z\\\"},\\\"endTime\\\":{\\\"S\\\":\\\"2022-01-01T10:00:40.561Z\\\"},\\\"id\\\":{\\\"S\\\":\\\"6e4bd304-446e-4678-8289-dasdasjkl\\\"},\\\"testStationName\\\":{\\\"S\\\":\\\"Rowe, Wunsch and Wisoky\\\"},\\\"activityType\\\":{\\\"S\\\":\\\"visit\\\"},\\\"activityDay\\\":{\\\"S\\\":\\\"2022-01-01\\\"},\\\"testerName\\\":{\\\"S\\\":\\\"231232132\\\"}},\\\"SequenceNumber\\\":\\\"1234\\\",\\\"SizeBytes\\\":704,\\\"StreamViewType\\\":\\\"NEW_AND_OLD_IMAGES\\\"},\\\"eventSourceARN\\\":\\\"arn:aws::eu--1::/cvs---//:32:37.491\\\"}\",\n  \"Timestamp\" : \"2024-03-27T14:27:25.926Z\",\n  \"SignatureVersion\" : \"1\",\n  \"Signature\" : \"+/+/+3//+//2f3y0TI+/+//---\"\n}";
+jest.mock("../../src/utils/sqsProcess.ts", () => ({
+  processRecord: mockProcessRecord
+}));
 
 describe("Retro Gen Function", () => {
   beforeAll(() => jest.setTimeout(60000));
@@ -77,9 +80,10 @@ describe("Retro Gen Function", () => {
 
     it("Should throw an error (generateATFReport fails)", async () => {
       ReportGenerationService.prototype.generateATFReport = jest.fn().mockRejectedValue(new Error("Oh no!"));
+      mockProcessRecord.mockReturnValueOnce("All good");
       expect.assertions(1);
       try {
-        await reportGen({ Records: [{ body: mockPayload }] }, ctx as any, () => {
+        await reportGen({ Records: [{ body: "test" }] }, ctx as any, () => {
           return;
         });
       } catch (e) {
@@ -89,9 +93,10 @@ describe("Retro Gen Function", () => {
     it("Should throw an error (bucket upload fails)", async () => {
       ReportGenerationService.prototype.generateATFReport = jest.fn().mockResolvedValue("Looking good");
       SendATFReport.prototype.sendATFReport = jest.fn().mockRejectedValue(new Error("Oh dear"));
+      mockProcessRecord.mockReturnValueOnce("All good");
       expect.assertions(1);
       try {
-        await reportGen({ Records: [{ body: mockPayload }] }, ctx as any, () => {
+        await reportGen({ Records: [{ body: "test" }] }, ctx as any, () => {
           return;
         });
       } catch (e) {
